align JOSE api to COSE api + move attestation
